#22 First version Warning: Database user need to be superuser to create extension 'hstore'

+++ sql/ws/ws_gw_trg_edit_node.sql
@@ -1,18 +1,18 @@
 /*
-This file is part of Giswater
+This file is part of Giswater 2.0
 The program is free software: you can redistribute it and/or modify it under the terms of the GNU General Public License as published by the Free Software Foundation, either version 3 of the License, or (at your option) any later version.
 This version of Giswater is provided by Giswater Association
 */
 
 
-   
-CREATE OR REPLACE FUNCTION SCHEMA_NAME.gw_trg_edit_node() RETURNS trigger LANGUAGE plpgsql AS $$
+CREATE OR REPLACE FUNCTION "SCHEMA_NAME"."gw_trg_edit_node"() RETURNS "pg_catalog"."trigger" AS $BODY$
 DECLARE 
     inp_table varchar;
     man_table varchar;
     v_sql varchar;
     old_nodetype varchar;
     new_nodetype varchar;
+
 
 BEGIN
 
@@ -26,37 +26,45 @@ BEGIN
             NEW.node_id:= (SELECT nextval('node_id_seq'));
         END IF;
 
+        -- Node type
+        IF (NEW.node_type IS NULL) THEN
+            IF ((SELECT COUNT(*) FROM node_type) = 0) THEN
+                RAISE EXCEPTION '[%]: There are no nodes types defined in the model, define at least one.', TG_NAME;
+            END IF;
+            NEW.node_type:= (SELECT id FROM node_type LIMIT 1);
+        END IF;
+
         -- Node Catalog ID
         IF (NEW.nodecat_id IS NULL) THEN
             IF ((SELECT COUNT(*) FROM cat_node) = 0) THEN
-                RAISE EXCEPTION 'There are no nodes catalog defined in the model, define at least one.';
-            END IF;   
+                RAISE EXCEPTION '[%]: There are no nodes catalog defined in the model, define at least one.', TG_NAME;
+            END IF;      
         END IF;
 
         -- Sector ID
         IF (NEW.sector_id IS NULL) THEN
             IF ((SELECT COUNT(*) FROM sector) = 0) THEN
-                RAISE EXCEPTION 'There are no sectors defined in the model, define at least one.';
+                RAISE EXCEPTION '[%]: There are no sectors defined in the model, define at least one.', TG_NAME;
             END IF;
-            NEW.sector_id := (SELECT sector_id FROM sector WHERE (NEW.the_geom @ sector.the_geom) LIMIT 1);
+            NEW.sector_id:= (SELECT sector_id FROM sector WHERE (NEW.the_geom @ sector.the_geom) LIMIT 1);
             IF (NEW.sector_id IS NULL) THEN
-                RAISE EXCEPTION 'Please take a look on your map and use the approach of the sectors!!!';
-            END IF;
+                RAISE EXCEPTION '[%]: Please take a look on your map and use the approach of the sectors!', TG_NAME;
+            END IF;            
         END IF;
         
         -- Dma ID
         IF (NEW.dma_id IS NULL) THEN
             IF ((SELECT COUNT(*) FROM dma) = 0) THEN
-                RAISE EXCEPTION 'There are no dma defined in the model, define at least one.';
+                RAISE EXCEPTION '[%]: There are no dma defined in the model, define at least one.', TG_NAME;
             END IF;
             NEW.dma_id := (SELECT dma_id FROM dma WHERE (NEW.the_geom @ dma.the_geom) LIMIT 1);
             IF (NEW.dma_id IS NULL) THEN
-                RAISE EXCEPTION 'Please take a look on your map and use the approach of the dma!!!';
-            END IF;
+                RAISE EXCEPTION '[%]: Please take a look on your map and use the approach of the dma!', TG_NAME;
+            END IF;            
         END IF;
         
-        -- FEATURE INSERT
-        INSERT INTO node VALUES (NEW.node_id, NEW.elevation, NEW."depth", NEW.nodecat_id, NEW.epa_type, NEW.sector_id, NEW."state", NEW.annotation, NEW."observ", NEW."comment",
+        -- FEATURE INSERT      
+        INSERT INTO node VALUES (NEW.node_id, NEW.elevation, NEW."depth", NEW.node_type, NEW.nodecat_id, NEW.epa_type, NEW.sector_id, NEW."state", NEW.annotation, NEW."observ", NEW."comment",
                                 NEW.dma_id, NEW.soilcat_id, NEW.category_type, NEW.fluid_type, NEW.location_type, NEW.workcat_id, NEW.buildercat_id, NEW.builtdate, 
                                 NEW.ownercat_id, NEW.adress_01, NEW.adress_02, NEW.adress_03, NEW.descript, NEW.rotation, NEW.link, NEW.verified, NEW.the_geom);
 
@@ -68,14 +76,17 @@ BEGIN
         ELSIF (NEW.epa_type = 'VALVE') THEN inp_table:= 'inp_valve';
         ELSIF (NEW.epa_type = 'SHORTPIPE') THEN inp_table:= 'inp_shortpipe';
         END IF;
-        v_sql:= 'INSERT INTO '||inp_table||' (node_id) VALUES ('||quote_literal(NEW.node_id)||')';
-        EXECUTE v_sql;
+        IF inp_table IS NOT NULL THEN        
+            v_sql:= 'INSERT INTO '||inp_table||' (node_id) VALUES ('||quote_literal(NEW.node_id)||')';
+            EXECUTE v_sql;
+        END IF;
 
         -- MANAGEMENT INSERT
         man_table:= (SELECT node_type.man_table FROM node_type JOIN cat_node ON (((node_type.id)::text = (cat_node.nodetype_id)::text)) WHERE cat_node.id=NEW.nodecat_id);
-        v_sql:= 'INSERT INTO '||man_table||' (node_id) VALUES ('||quote_literal(NEW.node_id)||')';
-        EXECUTE v_sql;
-            
+        IF man_table IS NOT NULL THEN
+            v_sql:= 'INSERT INTO '||man_table||' (node_id) VALUES ('||quote_literal(NEW.node_id)||')';
+            EXECUTE v_sql;
+        END IF;
         RETURN NEW;
 
 
@@ -96,8 +107,10 @@ BEGIN
             ELSIF (OLD.epa_type = 'PUMP') THEN
                 inp_table:= 'inp_pump';  
             END IF;
-            v_sql:= 'DELETE FROM '||inp_table||' WHERE node_id = '||quote_literal(OLD.node_id);
-            EXECUTE v_sql;
+            IF inp_table IS NOT NULL THEN
+                v_sql:= 'DELETE FROM '||inp_table||' WHERE node_id = '||quote_literal(OLD.node_id);
+                EXECUTE v_sql;
+            END IF;
 
             IF (NEW.epa_type = 'JUNCTION') THEN
                 inp_table:= 'inp_junction';   
@@ -112,21 +125,27 @@ BEGIN
             ELSIF (NEW.epa_type = 'PUMP') THEN
                 inp_table:= 'inp_pump';  
             END IF;
-            v_sql:= 'INSERT INTO '||inp_table||' (node_id) VALUES ('||quote_literal(NEW.node_id)||')';
-            EXECUTE v_sql;
+            IF inp_table IS NOT NULL THEN
+                v_sql:= 'INSERT INTO '||inp_table||' (node_id) VALUES ('||quote_literal(NEW.node_id)||')';
+                EXECUTE v_sql;
+            END IF;
 
         END IF;
 
-        IF (NEW.nodecat_id <> OLD.nodecat_id) THEN  
-            old_nodetype:= (SELECT node_type.type FROM node_type JOIN cat_node ON (((node_type.id)::text = (cat_node.nodetype_id)::text)) WHERE cat_node.id=OLD.nodecat_id)::text;
-            new_nodetype:= (SELECT node_type.type FROM node_type JOIN cat_node ON (((node_type.id)::text = (cat_node.nodetype_id)::text)) WHERE cat_node.id=NEW.nodecat_id)::text;
-            IF (quote_literal(old_nodetype)::text <> quote_literal(new_nodetype)::text) THEN
-                RAISE EXCEPTION 'Change node catalog is forbidden. The new node catalog is not included on the same type (node_type.type) of the old node catalog';
+        IF (NEW.node_type <> OLD.node_type) THEN  
+                RAISE EXCEPTION '[%]: Change node catalog is forbidden. The new node catalog is not included on the same type (node_type.type) of the old node catalog',TG_NAME;
+        END IF;
+
+        IF (OLD.nodecat_id IS NOT NULL) AND (NEW.nodecat_id <> OLD.nodecat_id) THEN  
+            old_nodetype:= (SELECT node_type.type FROM node_type JOIN cat_node ON (((node_type.id) = (cat_node.nodetype_id))) WHERE cat_node.id=OLD.nodecat_id);
+            new_nodetype:= (SELECT node_type.type FROM node_type JOIN cat_node ON (((node_type.id) = (cat_node.nodetype_id))) WHERE cat_node.id=NEW.nodecat_id);
+            IF (quote_literal(old_nodetype) <> quote_literal(new_nodetype)) THEN
+                RAISE EXCEPTION '[%]: Change node catalog is forbidden. The new node catalog is not included on the same type (node_type.type) of the old node catalog', TG_NAME;
             END IF;
         END IF;
 
         UPDATE node 
-        SET node_id=NEW.node_id, elevation=NEW.elevation, "depth"=NEW."depth", nodecat_id=NEW.nodecat_id, epa_type=NEW.epa_type, sector_id=NEW.sector_id, "state"=NEW."state", 
+        SET node_id=NEW.node_id, elevation=NEW.elevation, "depth"=NEW."depth", node_type=NEW.node_type, nodecat_id=NEW.nodecat_id, epa_type=NEW.epa_type, sector_id=NEW.sector_id, "state"=NEW."state", 
             annotation=NEW.annotation, "observ"=NEW."observ", "comment"=NEW."comment", dma_id=NEW.dma_id, soilcat_id=NEW.soilcat_id, category_type=NEW.category_type, 
             fluid_type=NEW.fluid_type, location_type=NEW.location_type, workcat_id=NEW.workcat_id, buildercat_id=NEW.buildercat_id, builtdate=NEW.builtdate,
             ownercat_id=NEW.ownercat_id, adress_01=NEW.adress_01, adress_02=NEW.adress_02, adress_03=NEW.adress_03, descript=NEW.descript,
@@ -144,11 +163,12 @@ BEGIN
     END IF;
 
 END;
-$$;
+$BODY$
+  LANGUAGE 'plpgsql' VOLATILE COST 100
+;
 
 
 
-CREATE TRIGGER gw_trg_edit_node INSTEAD OF INSERT OR DELETE OR UPDATE ON SCHEMA_NAME.v_edit_node
-FOR EACH ROW EXECUTE PROCEDURE SCHEMA_NAME.gw_trg_edit_node();
-
+CREATE TRIGGER gw_trg_edit_node INSTEAD OF INSERT OR DELETE OR UPDATE ON "SCHEMA_NAME".v_edit_node
+FOR EACH ROW EXECUTE PROCEDURE "SCHEMA_NAME".gw_trg_edit_node();
